Move config updating for peer endpoint overrides to
