+++ client-core/src/lib.rs
@@ -7,6 +7,7 @@ pub use innernet_shared::{
     Peer, WrappedIoError, DEFAULT_HOSTS_PATH,
 };
 pub use wireguard_control::Backend;
+use wireguard_control::{DeviceUpdate, Key, PeerConfigBuilder};
 
 use anyhow::Error;
 use innernet_shared::wg;
@@ -39,6 +40,39 @@ pub fn set_listen_port(
     config.interface.listen_port = listen_port;
     config.save(config_dir, interface)?;
     log::info!("the config file is updated");
+
+    Ok(())
+}
+
+pub fn set_endpoint_override_for_peer(
+    network_backend: Backend,
+    config_dir: &Path,
+    interface: &interface::InterfaceName,
+    config: &mut interface::InterfaceConfig,
+    peer: &Peer,
+    endpoint: Endpoint,
+) -> Result<(), Error> {
+    config.set_endpoint_override_for_peer(peer.ip, endpoint.clone());
+    config.save(config_dir, interface)?;
+
+    let socket_addr = endpoint.resolve()?;
+    let peer_pub_key = Key::from_base64(&peer.public_key).unwrap();
+
+    DeviceUpdate::new()
+        .add_peer(PeerConfigBuilder::new(&peer_pub_key).set_endpoint(socket_addr))
+        .apply(interface, network_backend)?;
+
+    Ok(())
+}
+
+pub fn unset_endpoint_override_for_peer(
+    config_dir: &Path,
+    interface: &interface::InterfaceName,
+    config: &mut interface::InterfaceConfig,
+    peer: &Peer,
+) -> Result<(), Error> {
+    config.unset_endpoint_override_for_peer(peer.ip);
+    config.save(config_dir, interface)?;
 
     Ok(())
 }
